% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-effects.R
\name{.ame_build_decomp}
\alias{.ame_build_decomp}
\title{Build the AME decomposition table}
\usage{
.ame_build_decomp(ext_summ, int_summ, total_summ, labels, idx)
}
\arguments{
\item{ext_summ}{Data frame: extensive summary from .ame_summarize.}

\item{int_summ}{Data frame: intensive summary from .ame_summarize.}

\item{total_summ}{Data frame: total summary from .ame_summarize.}

\item{labels}{Character vector of length P.}

\item{idx}{Integer vector: indices of non-intercept covariates.}
}
\value{
Data frame with 13 columns.
}
\description{
Build the AME decomposition table
}
\keyword{internal}
