% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdlebb-package.R
\docType{package}
\name{hurdlebb-package}
\alias{hurdlebb}
\alias{hurdlebb-package}
\title{hurdlebb: Hurdle Beta-Binomial Models for Bounded Discrete Proportions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fits two-part hurdle beta-binomial models for bounded discrete proportions with structural zeros. Supports survey-weighted pseudo-posterior inference, state-varying coefficients, sandwich standard errors, and marginal effect decomposition via 'CmdStan'. Designed for analysis of complex survey data where a proportion of units report zero counts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://joonho112.github.io/hurdlebb/}
  \item \url{https://github.com/joonho112/hurdlebb}
  \item Report bugs at \url{https://github.com/joonho112/hurdlebb/issues}
}

}
\author{
\strong{Maintainer}: JoonHo Lee \email{jlee233@ua.edu} (\href{https://orcid.org/0009-0006-4019-8703}{ORCID})

}
\keyword{internal}
