% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbb-fit-class.R
\name{is.hbb_fit}
\alias{is.hbb_fit}
\title{Test if an Object is an hbb_fit}
\usage{
is.hbb_fit(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical scalar.
}
\description{
Returns \code{TRUE} if \code{x} inherits from class \code{"hbb_fit"}, \code{FALSE}
otherwise. This is the recommended way to check whether an object
was produced by \code{\link[=hbb]{hbb()}}.
}
\examples{
is.hbb_fit(1)          # FALSE
is.hbb_fit(list())     # FALSE

\dontrun{
fit <- hbb(y | trials(n_trial) ~ poverty, data = my_data)
is.hbb_fit(fit)        # TRUE
}

}
\seealso{
\code{\link[=hbb]{hbb()}} for fitting models.

Other fitting: 
\code{\link{hbb}()},
\code{\link{print.hbb_fit}()}
}
\concept{fitting}
