% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.hbb_fit}
\alias{nobs.hbb_fit}
\title{Number of Observations in an hbb_fit}
\usage{
\method{nobs}{hbb_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"hbb_fit"} returned by
\code{\link{hbb}}.}

\item{...}{Currently unused; included for S3 method consistency.}
}
\value{
An integer scalar giving the number of observations.
}
\description{
Extracts the number of provider-level observations \eqn{N} used to
fit the hurdle Beta-Binomial model.
}
\examples{
\dontrun{
fit <- hbb(y | trials(n_trial) ~ poverty + urban, data = my_data)
nobs(fit)
}

}
\seealso{
\code{\link{hbb}}, \code{\link{coef.hbb_fit}},
\code{\link{summary.hbb_fit}}
}
